third_party/WebKit/LayoutTests/fast/dom/HTMLDocument/clone-node-expected.txt

Tests cloneNode for HTMLDocument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.cloneNode(false).__proto__ is HTMLDocument.prototype
PASS className(document.cloneNode(false)) is "HTMLDocument"
PASS document.cloneNode(true).title is document.title
PASS document.cloneNode(true).compatMode is "BackCompat"
PASS document.cloneNode(true).contentType is "text/html"
PASS doc.cloneNode(false).__proto__ is HTMLDocument.prototype
PASS className(doc.cloneNode(false)) is "HTMLDocument"
PASS doc.cloneNode(true).title is doc.title
PASS doc.cloneNode(true).compatMode is "CSS1Compat"
PASS doc.cloneNode(true).contentType is "text/html"
PASS successfullyParsed is true

TEST COMPLETE
